refactor(core): move file-IO config/theme wrappers out of the pure core into app/
Phase 6a of the pure-core/portable-support split. The core's last three
file-IO wrappers move to app/, leaving DanTermCore free of filesystem reads
and of any unmarked NSHomeDirectory:

- DanTermConfigParser.configFilePath() (NSHomeDirectory-based) -> new app-level
  enum DanTermConfigPaths.configFilePath(). Repointed its five app callers
  (AppRuntime x2, AppDelegate, PreferencesPanel, and GhosttyApp's DanTerm
  overlay loader). GhosttyApp's own configFilePath() -- the Ghostty overlay
  path, a different symbol -- is untouched.
- DanTermConfigParser.loadFromDisk() (FileManager read) -> app-side extension,
  keeping the name so its two AppRuntime callers are unchanged; it calls the
  new DanTermConfigPaths.configFilePath().
- ThemeColorParser.parse(themeFileAt:) (FileManager read) -> app-side extension,
  keeping the name so its ThemeCatalog caller is unchanged.

The pure parser/writer they wrap (parse(content:), DanTermConfigWriter,
parse(themeContent:), ThemeColorHex) stay in core as pure string transforms.

No public/package annotations: the moved helpers are app-internal, reached
same-module via the app/DanTermCore symlink. The lint that enforces the now-IO-
free invariant lands next; this commit is green on the existing Cocoa-only lint.

CLI surface unchanged -> no integrations/danterm/SKILL.md edit.

Green: core 1055/37, just build, just test-ui 12/12, pure lint + self-test.
